mvoe abstract class into its own file
