+++ app/app/javascript/adapters/pinwheel.ts
@@ -1,40 +1,10 @@
 import { fetchToken, trackUserAction } from "../utilities/api";
 import loadScript from 'load-script';
 import { getDocumentLocale } from "../utilities/getDocumentLocale";
+import { IncomeDataAdapter } from "./IncomeDataAdapter";
 
 declare global {
     var Pinwheel: any;
-}
-
-abstract class IncomeDataAdapter {
-    successCallback?: Function;
-    exitCallback?: Function;
-    
-    abstract open(responseType : string, id : string, name : string, isDefaultOption : boolean): void;
-    abstract onEvent(eventName : string, eventPayload : any): void;
-    abstract load();
-
-    constructor(args : {
-        onSuccess?: Function;
-        onExit?: Function;
-    } = { onSuccess: () => {}}) {
-        this.successCallback = args.onSuccess;
-        this.exitCallback = args.onExit;
-
-        this.load();
-    }
-
-    async onExit(eventPayload: any) {
-        if (this.exitCallback) {
-            this.exitCallback()
-        }
-    }
-
-    async onSuccess(eventPayload: any) {
-        if (this.successCallback) {
-            this.successCallback()
-        }
-    }
 }
 
 export default class PinwheelIncomeDataAdapter extends IncomeDataAdapter {
